+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - node - after node-v14 is deprecated, remove shell-code `export "NODE_OPTIONS=--unhandled-rejections=strict"`.
 
 # v2021.10.1-beta
+- doc - add api-documentation
+- jslint - relax regexp-warning against using 'space'
 - ci - rename dir .build/ to .artifact/
 - fs - merge file asset_codemirror_rollup.css into index.html
 - fs - merge file function.html into help.html